We use `localStorage` to synchronize changes across tabs. We write the
root hash to `localStorage` which triggers `window.onstorage` in other
tabs running on the same domain. We then run `_checkChange` to run all
subscriptions as needed.

Fixes https://github.com/rocicorp/repc/issues/163