@phritz 2020-09-22
Might be a good idea to address https://github.com/rocicorp/repc/issues/180 after you land this, and ensure that our current customers are aware of this behavior. There is a right way and a wrong way for the ui to behave given this behavior so imho worthwhile ensuring everyone is clear on that.