@arv 2020-09-22
Perf results:

master branch

```
Running benchmarks please wait...
populate 1024x1000 (clean) | Median: 761 | Throughput: 1.28 MB/s
populate 1024x1000 (dirty) | Median: 1282 | Throughput: 780.03 KB/s
scan 1024x1000 | Median: 94 | Throughput: 10.39 MB/s
scan 1024x5000 | Median: 448 | Throughput: 10.90 MB/s
Done!
Running benchmarks please wait...
populate 1024x1000 (clean) | Median: 699 | Throughput: 1.40 MB/s
populate 1024x1000 (dirty) | Median: 1261 | Throughput: 793.02 KB/s
scan 1024x1000 | Median: 111 | Throughput: 8.80 MB/s
scan 1024x5000 | Median: 609 | Throughput: 8.02 MB/s
Done!
```

This branch
```
Running benchmarks please wait...
populate 1024x1000 (clean) | Median: 670 | Throughput: 1.46 MB/s
populate 1024x1000 (dirty) | Median: 1274 | Throughput: 784.93 KB/s
scan 1024x1000 | Median: 82 | Throughput: 11.91 MB/s
scan 1024x5000 | Median: 421 | Throughput: 11.60 MB/s
Done!
Running benchmarks please wait...
populate 1024x1000 (clean) | Median: 748 | Throughput: 1.31 MB/s
populate 1024x1000 (dirty) | Median: 1361 | Throughput: 734.75 KB/s
scan 1024x1000 | Median: 129 | Throughput: 7.57 MB/s
scan 1024x5000 | Median: 472 | Throughput: 10.34 MB/s
Done!
```

Not sure how it got faster ;-)